pyproject.toml
[tool.poetry]
name = "tencentcloud-sdk-python-ext"
version = "0.1.0"
description = "腾讯云Python Sdk 插件"
authors = ["thomas <thomas.mingyu.guo@gmail.com>"]
readme = "README.md"
packages = [{include = "tcc_ext"}]

repository = "https://github.com/mingyugg/tencentcloud-sdk-python-ext"

[tool.poetry.dependencies]
python = "^3.10"
tencentcloud-sdk-python-common = "^3.0.1023"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
